Refactor TestimonialCard into own component file
